Create Best_Loyalty_Program_Results
The supermarket has members and normal customers. This script of code answers the question, which branch has the best results in the loyalty program? We want the branches, the amount of members at each store, and the total revenue from each store. We also want to filter out all transactions that are conducted with normal customers to assess the numbers from members.