Create Popular_Product_Line_For_Women
In this query, we count the amount of transactions for each product line and the total amount of items purchased in each product line. We also use the WHERE clause to find transactions made by females. We order the query by the amount of items purchased from most to least purchased.