Create Gross_Income_vs_Payment_Method
In this script, we use the supermarket dataset answer the question, does gross income depend on payment method? We find this by using the AVG function to get the average of gross income rounded to 2 decimal digits. We group the query by payment methods which are e-wallet, cash, and credit card.