Create Highest_Income_Category
In this script, we use the supermarket dataset to answer the following question, which product category generates the highest income? We will use the SUM function and the ROUND function. We get the sum of all revenue for each product line, rounded to two decimal digits. We get the cost of goods by finding the sum of the cost of goods for each product line. We get the gross income by getting the sum of gross income for each product line rounded to two digits. We group by each product line. Then we order by gross income from highest to lowest.